Move Jupiter methods to new service and lazy load
